backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
langchain==0.0.350
langchain-openai==0.0.2
pinecone-client==2.2.4
pypdf2==3.0.1
python-dotenv==1.0.0
pydantic==1.10.13

